+++ messenger-tfs/src/main/java/com/spinoza/messenger_tfs/domain/model/Message.kt
@@ -2,15 +2,15 @@ package com.spinoza.messenger_tfs.domain.model
 
 data class Message(
     val date: MessageDate,
-    val userId: Int,
+    val userId: Long,
     val name: String,
     val content: String,
     val avatarResId: Int,
     val reactions: Map<String, ReactionParam>,
     val isIconAddVisible: Boolean,
-    val id: Int = UNDEFINED_ID,
+    val id: Long = UNDEFINED_ID,
 ) {
     companion object {
-        const val UNDEFINED_ID = -1
+        const val UNDEFINED_ID = -1L
     }
 }
