+++ messenger-tfs/src/main/java/com/spinoza/messenger_tfs/domain/model/MessagePosition.kt
@@ -2,7 +2,7 @@ package com.spinoza.messenger_tfs.domain.model
 
 class MessagePosition(
     val type: Type = Type.UNDEFINED,
-    val messageId: Int = Message.UNDEFINED_ID,
+    val messageId: Long = Message.UNDEFINED_ID,
 ) {
     enum class Type {
         UNDEFINED,
