+++ messenger-tfs/src/main/java/com/spinoza/messenger_tfs/presentation/adapter/delegate/message/CompanionMessageDelegateItem.kt
@@ -18,7 +18,7 @@ class CompanionMessageDelegateItem(
         return value
     }
 
-    override fun id(): Int {
+    override fun id(): Long {
         return value.id
     }
