+++ messenger-tfs/src/main/java/com/spinoza/messenger_tfs/presentation/adapter/delegate/message/UserMessageDelegateItem.kt
@@ -18,7 +18,7 @@ class UserMessageDelegateItem(
         return value
     }
 
-    override fun id(): Int {
+    override fun id(): Long {
         return value.id
     }
